+++ test/app.ts
@@ -3,30 +3,12 @@ import { Cocobase } from "../src/index.ts";
 import { ref } from "process";
 
 const db = new Cocobase({
+  apiKey:"WOPaOrZQpXumrLizEYyZNhQ3EPr_SElu5uTI-CJH",
+  projectId:"d52bba2b-0ff8-4663-b7cf-fe88144ad0d1",
 });
 
-// todo test updateDocuments (batch request)
-// todo test createDocumentWithFiles
-// todo test updateDocumentWithFiles
-// todo create user with files
-// todo implement aggregation queries
-// *STUFF TO TEST*
-// /collections/{collection_id}/query/documents/group-by
-// Group By Field
-// GET
-// /collections/{id}/query/schema
-// Get Collection Schema
-// GET
-// /collections/{id}/export
-// Export Collection
 async function run() {
-  const users = await db.auth.listUsers({
-    filters: {
-      "data.referral_balance_gte":20
-    },
-    limit:2
-  });
+  await db.auth.login("demo@gmail.com","demodemo")
 
-  console.log(JSON.stringify(users,null,2));
 }
 run();